requirements.txt
alembic==1.4.2
amqp==2.6.0
bcrypt==3.1.7
billiard==3.6.3.0
blinker==1.4
celery==4.4.6
cffi==1.14.0
click==7.1.2
defusedxml==0.6.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cache==0.13.1
Flask-Celery-Helper==1.1.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-OAuth==0.12
Flask-OpenID==1.2.5
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
future==0.18.2
httplib2==0.18.1
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.11
Mako==1.1.3
MarkupSafe==1.1.1
oauth2==1.9.0.post1
pycparser==2.20
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.2.0
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
vine==1.3.0
Werkzeug==0.15.2
WTForms==2.3.1
zipp==3.1.0
